% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{Example Multivariate Data}
\format{A \code{data.frame}s with 32 variables}
\usage{
dat
}
\description{
A data with potentially high skewness and kurtosis
}
\keyword{datasets}
