% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoevd.R
\name{hoevd}
\alias{hoevd}
\title{Higher Order Eigen Value Decomposition of A Symmetric Tensor}
\usage{
hoevd(x, rank = p)
}
\arguments{
\item{x}{the symmetric tensorian stored as an array}

\item{rank}{rank of the decomposition}
}
\value{
A list with eliments
\item{u}{Estimated eigen vectors}
\item{z}{Estimated core tensor}
}
\description{
Decomposes a higher order symmetric tensors (e.g. higher order centered and
standardized tensorian moments) using HOEVD. It is a special case of HOSVD.
}
\details{
x = Z \code{*} [u,..u]
}
\examples{
data(dat)
moment_center_sd_3 = tensor_moment(Y = dat, center = TRUE, standardize = TRUE, p = 3)$moment
decomp_3 = hoevd(x = moment_center_sd_3, rank = 32)

}
