% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rob_TPM.R
\name{Rob_TPM}
\alias{Rob_TPM}
\title{Robust Power Tensor Method for Decomposition of A Symmetric Moment Tenosr}
\usage{
Rob_TPM(Y, center = TRUE, standardize = TRUE, L = 10, N = 10,
  order = c(3, 4), rank = ncol(Y), tol = 1e-06)
}
\arguments{
\item{Y}{the symmetric tensorian stored as an array}

\item{center}{column-center the data first, default is \code{TRUE}.}

\item{standardize}{standardize the multivariate data, i.e. convert it to the left singular matrix}

\item{L}{number of outloop, i.e. number of power tensor iterations}

\item{N}{number of inner loop, i.e. steps for each power tensor iteration to converge}

\item{order}{order of the decomposition, 3 or 4}

\item{rank}{rank of the decomposition}

\item{tol}{tolerance for the inner loop, i.e power tensor iteration, default to be 1e-06}
}
\value{
A list with eliments
\item{eigenv}{Estimated eigen vectors, sorted by based on eigenl}
\item{eigenl}{Estimated eigen values, sorted from largest to smallest}
}
\description{
Decomposes a higher order symmetric tensors (e.g. higher order centered and
standardized tensorian moments) using Robut Power Tensor Method (RPTM) based on the vectorization
provided by Di et al.
}
\details{
See Di et al. 2018. This function does not contruct tensor first. It can only be used directly to the
data matrix, and it produces decomposition of the (standardized) moment tensors
}
\examples{
data(dat)
rob_3 = Rob_TPM(Y = dat, order = 3, rank = 10, tol = 1e-6)

}
\references{
Di et al.

A Anandkumar et al. Tensor decompositions for learning latent variable models, 2012.
}
