% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_cumulant.R
\name{tensor_cumulant}
\alias{tensor_cumulant}
\title{Construct Higher Order Tensorian Cumulants}
\usage{
tensor_cumulant(Y, p)
}
\arguments{
\item{Y}{data.frame or matrix of multivariate data of dimension n*p}

\item{p}{what is the order to construct}
}
\value{
3rd or 4th order cumulant tensors
}
\description{
3rd and 4th order tensorian cumulants
}
\examples{
data(dat)
tensor_3 = tensor_cumulant(Y = dat, p = 3)
}
