% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_moment.R
\name{tensor_moment}
\alias{tensor_moment}
\title{Construct Higher Order Tensorian Moments}
\usage{
tensor_moment(Y, center = TRUE, standardize = TRUE, p = c(3, 4))
}
\arguments{
\item{Y}{data.frame or matrix of multivariate data of dimension n*p}

\item{center}{column-center the data first, default is \code{TRUE}.}

\item{standardize}{standardize the multivariate data, i.e. convert it to the left singular matrix}

\item{p}{what is the order to construct}
}
\value{
3rd or 4th order moments
}
\description{
3rd and 4th order (centered and standardized) tensorian moments
}
\details{
Suppose Y = USV' is the SVD, standardization is done by taking U only
}
\examples{
data(dat)
moment_center_sd_3 = tensor_moment(Y = dat, center = TRUE, standardize = TRUE, p = 3)
}
